requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0
executing==1.2.0
gunicorn==20.1.0
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2==2.9.5
pure-eval==0.2.2
Pygments==2.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.5.0
tzdata==2023.3
wcwidth==0.2.6
whitenoise==6.4.0
